Richard/cmake

+++ src/lib/dglib/lib/DgRFNetwork.cpp
@@ -22,10 +22,10 @@
 //
 ////////////////////////////////////////////////////////////////////////////////
 
-#include "DgBase.h"
-#include "DgRFBase.h"
-#include "DgLocation.h"
-#include "DgSeriesConverter.h"
+#include <dglib/DgBase.h>
+#include <dglib/DgRFBase.h>
+#include <dglib/DgLocation.h>
+#include <dglib/DgSeriesConverter.h>
 
 ////////////////////////////////////////////////////////////////////////////////
 DgRFNetwork::~DgRFNetwork (void)
@@ -41,7 +41,7 @@ DgRFNetwork::~DgRFNetwork (void)
       matrix_[i].clear();
    }
 
-   for (int i = 0; i < frames_.size(); i++) {
+   for (size_t i = 0; i < frames_.size(); i++) {
       delete frames_[i];
       frames_[i] = 0;
    }
@@ -58,8 +58,8 @@ DgRFNetwork::existsConverter (const DgRFBase& fromFrame,
 } // bool DgRFNetwork::existsConverter
 
 ////////////////////////////////////////////////////////////////////////////////
-const DgConverterBase* 
-DgRFNetwork::getConverter (const DgLocation& fromLoc, 
+const DgConverterBase*
+DgRFNetwork::getConverter (const DgLocation& fromLoc,
                                  const DgRFBase& toFrame) const
 {
    return getConverter(fromLoc.rf(), toFrame);
@@ -67,8 +67,8 @@ DgRFNetwork::getConverter (const DgLocation& fromLoc,
 } // const DgConverter* DgRFNetwork::getConverter
 
 ////////////////////////////////////////////////////////////////////////////////
-const DgConverterBase* 
-DgRFNetwork::getConverter (const DgRFBase& fromFrame, 
+const DgConverterBase*
+DgRFNetwork::getConverter (const DgRFBase& fromFrame,
                                  const DgRFBase& toFrame) const
 {
    // check for network match
@@ -82,10 +82,10 @@ DgRFNetwork::getConverter (const DgRFBase& fromFrame,
 
    // check for existing converter
 
-   if (!matrix_[fromFrame.id()][toFrame.id()]) 
+   if (!matrix_[fromFrame.id()][toFrame.id()])
    {
       // attempt to generate a passthrough converter
-      
+
       if (!fromFrame.connectTo() || !toFrame.connectFrom())
       {
          report("DgRFNetwork::getConverter() frames not connected: " +
@@ -93,9 +93,9 @@ DgRFNetwork::getConverter (const DgRFBase& fromFrame,
                 DgBase::Fatal);
          return 0;
       }
-      
+
       const_cast<DgRFNetwork*>(this)->matrix_
-                           [fromFrame.id()][toFrame.id()] = 
+                           [fromFrame.id()][toFrame.id()] =
                                 new DgSeriesConverter(fromFrame, toFrame);
    }
 
@@ -131,7 +131,7 @@ DgRFNetwork::reserve (const size_t& capacity)
    frames_.reserve(capacity);
    matrix_.reserve(capacity);
 
-   for (unsigned long i = 0; i < matrix_.size(); i++) 
+   for (unsigned long i = 0; i < matrix_.size(); i++)
     matrix_[i].reserve(capacity);
 
 } // void DgRFNetwork::reserve
@@ -145,22 +145,22 @@ DgRFNetwork::generateId (DgRFBase* frame)
    frames_.resize(newSize, 0);
    matrix_.resize(newSize);
 
-   for (unsigned long long int i = 0; i < newSize; i++) 
+   for (unsigned long long int i = 0; i < newSize; i++)
     matrix_[i].resize(newSize, 0);
-   
+
    frames_[nextId_] = frame;
-   
+
    matrix_[nextId_][nextId_] = new DgIdentityConverter(*frame);
-   
-   if (nextId_ == 0) 
+
+   if (nextId_ == 0)
     frame->connectTo_ = frame->connectFrom_ = frame; // ground
 
-   if (nextId_ == (int) capacity()) 
+   if (nextId_ == (int) capacity())
     reserve(capacity() + chunkSize());
-   
+
    nextId_++;
-   
+
    return (nextId_ - 1);
-} 
+}
 
 ////////////////////////////////////////////////////////////////////////////////
